Add second file test_get_book_data_methods to test those functions, rename test.py to test_clean_search_methods.py to break up testing
